refactor(dashboard): move lens chart presentation helpers to lenscommon
