+++ internal/kibana/dashboard/models_lens_resolver.go
@@ -1,0 +1,48 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package dashboard
+
+import (
+	"github.com/elastic/terraform-provider-elasticstack/generated/kbapi"
+	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/lenscommon"
+	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/models"
+)
+
+type chartPresentationResolver struct {
+	dashboard *models.DashboardModel
+}
+
+// lensChartResolver implements lenscommon.Resolver using dashboard-level defaults from the enclosing dashboard model.
+func lensChartResolver(dashboard *models.DashboardModel) lenscommon.Resolver {
+	return &chartPresentationResolver{dashboard: dashboard}
+}
+
+func (r *chartPresentationResolver) ResolveChartTimeRange(chartLevel *models.TimeRangeModel) kbapi.KbnEsQueryServerTimeRangeSchema {
+	return resolveChartTimeRange(r.dashboard, chartLevel)
+}
+
+func (r *chartPresentationResolver) DashboardLensComparableTimeRange() (kbapi.KbnEsQueryServerTimeRangeSchema, bool) {
+	return dashboardLensComparableTimeRange(r.dashboard)
+}
+
+func dashboardLensComparableTimeRange(dashboard *models.DashboardModel) (kbapi.KbnEsQueryServerTimeRangeSchema, bool) {
+	if dashboard == nil || dashboard.TimeRange == nil {
+		return kbapi.KbnEsQueryServerTimeRangeSchema{}, false
+	}
+	return timeRangeModelToAPI(dashboard.TimeRange), true
+}
